# test: perf: perf_simple_query: collect allocation and task statistics